+++ packages/gatsby-plugin-image/compat.browser.modern.js
@@ -1,0 +1,1 @@
+export * from "./dist/gatsby-image.compat.browser.modern"