+++ packages/gatsby-plugin-page-creator/src/derive-path.ts
@@ -1,5 +1,6 @@
 import _ from "lodash"
 import slugify from "@sindresorhus/slugify"
+import { Reporter } from "gatsby"
 import {
   compose,
   removeFileExtension,
@@ -7,7 +8,8 @@ import {
   extractAllCollectionSegments,
   switchToPeriodDelimiters,
 } from "./path-utils"
-import { Reporter } from "gatsby"
+
+const doubleForwardSlashes = /\/\/+/g
 
 // Generates the path for the page from the file path
 // product/{Product.id}.js => /product/:id, pulls from nodes.id
@@ -26,7 +28,7 @@ export function derivePath(
 
   // 3.  For each slug parts get the actual value from the node data
   slugParts.forEach(slugPart => {
-    // 3.a. this transforms foo__bar into foo.bar
+    // 3.a.  this transforms foo__bar into foo.bar
     const key = compose(
       extractFieldWithoutUnion,
       switchToPeriodDelimiters
@@ -51,24 +53,19 @@ export function derivePath(
     pathWithoutExtension = pathWithoutExtension.replace(slugPart, value)
   })
 
-  return pathWithoutExtension
+  // 4.  Remove double forward slashes that could occur in the final URL
+  const derivedPath = pathWithoutExtension.replace(doubleForwardSlashes, `/`)
+
+  return derivedPath
 }
 
+// If the node value is meant to be a slug, like `foo/bar`, the slugify
+// function will remove the slashes. This is a hack to make sure the slashes
+// stick around in the final url structuring
 function safeSlugify(nodeValue: string): string {
-  // If the node value is meant to be a slug, like `foo/bar`, the slugify
-  // function will remove the slashes. This is a hack to make sure the slashes
-  // stick around in the final url structuring
-  const SLASH_PRESERVING_STATIC_KEY = `-replaced-`
+  // The incoming GraphQL data can also be a number
+  const input = String(nodeValue)
+  const tempArr = input.split(`/`)
 
-  const replaceSlashesValue = (nodeValue + ``).replace(
-    /\//g,
-    SLASH_PRESERVING_STATIC_KEY
-  )
-
-  const slugifiedWithoutSlashesValue = slugify(replaceSlashesValue)
-
-  return slugifiedWithoutSlashesValue.replace(
-    new RegExp(SLASH_PRESERVING_STATIC_KEY, `g`),
-    `/`
-  )
+  return tempArr.map(v => slugify(v)).join(`/`)
 }