+++ packages/gatsby-plugin-typescript/src/gatsby-node.js
@@ -35,6 +35,21 @@ function onCreateWebpackConfig({ actions, loaders }) {
   })
 }
 
+if (process.env.GATSBY_EXPERIMENTAL_PLUGIN_OPTION_VALIDATION) {
+  exports.pluginOptionsSchema = ({ Joi }) =>
+    Joi.object({
+      isTSX: Joi.boolean().description(`Enables jsx parsing.`).default(true),
+      jsxPragma: Joi.string()
+        .description(
+          `Replace the function used when compiling JSX expressions.`
+        )
+        .default(`React`),
+      allExtensions: Joi.boolean()
+        .description(`Indicates that every file should be parsed as TS or TSX.`)
+        .default(false),
+    })
+}
+
 exports.resolvableExtensions = resolvableExtensions
 exports.onCreateBabelConfig = onCreateBabelConfig
 exports.onCreateWebpackConfig = onCreateWebpackConfig