+++ packages/gatsby/src/utils/api-node-docs.ts
@@ -428,3 +428,10 @@ export const onPreExtractQueries = true
  * }
  */
 export const onCreateDevServer = true
+
+/**
+ * Called during `gatsby develop` bootstrap to get and validate a plugins options schema
+ * @param {Joi} $0.Joi The instance of Joi to define the schema with
+ *
+ */
+export const pluginOptionsSchema = true